Refactoring.

+++ chat_channel/views.py
@@ -39,6 +39,7 @@ class ChatChannelView(generics.CreateAPIView,
         `workspace_hashed_value`를 입력하면 해당 workspace의 `chat_channel`을 추가합니다.
         `name`에는 만들 채널의 이름을 넣으십시오.
         채널을 생성하면 `members`안에는 자동으로 만든 사람의 정보가 포함 됩니다.
+        채널 만든 사람은 `admins`안에 역시 자동으로 들어갑니다.
         """
         if request.data.get('name', None) is None:
             return Response({'msg': 'name field is not filled.'}, status=status.HTTP_400_BAD_REQUEST)
@@ -69,8 +70,12 @@ class ChatChannelView(generics.CreateAPIView,
         `ChatChannel`의 설명 문구를 바꾸기 위한 엔드포인트 입니다.
         이걸로 `ChatChannel`의 이름이나 `members`는 못바꿉니다.
         body의 `name`은 설명을 바꿀 `ChatChannel`의 이름입니다.
+        `admins`안에 포함되지 않은 유저는 401이 나옵니다.
         """
         chat_channel = self.get_queryset().get(name__exact=request.data.get('name', None))
+
+        if request.user not in chat_channel.admins.all():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
 
         chat_channel.description = request.data.get('description', None)
         chat_channel.save()
@@ -94,22 +99,31 @@ class ChatChannelUpdateDeleteView(generics.UpdateAPIView,
 
     def delete(self, request: Request, *args, **kwargs):
         chat_channel = self.get_queryset()
+
+        if request.user not in chat_channel.admins.all():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         chat_channel.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
+    @swagger_auto_schema(operation_id="channel_name_update")
     def patch(self, request: Request, *args, **kwargs):
         """
         `ChatChannel`의 이름을 바꿀때 사용합니다.
         path에 `channel_name`은 대상의 이름, body에 `name`은 새로 바꿀 이름입니다.
         """
         chat_channel: ChatChannel = self.get_queryset()
+
+        if request.user not in chat_channel.admins.all():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         chat_channel.name = request.data.get('name', None)
         chat_channel.save()
         serializer = self.get_serializer(chat_channel)
         return Response(serializer.data)
 
 
-class ChatChannelAddMembersView(generics.UpdateAPIView):
+class ChatChannelAddMembersView(generics.CreateAPIView):
     queryset = ChatChannel.objects.all()
     http_method_names = ['post']
     permission_classes = [permissions.IsAuthenticated]
@@ -125,6 +139,7 @@ class ChatChannelAddMembersView(generics.UpdateAPIView):
         """
         채널에 인원을 추가 하고 싶을 때 쓰는 엔드포인트 입니다.
         배열을 쓰면 여러명을 추가 할 수 있습니다.
+        `admins`에 포함되지 않은 유저는 401이 반환 됩니다.
         """
         chat_channel: ChatChannel = self.get_queryset()
         users = []
@@ -135,14 +150,17 @@ class ChatChannelAddMembersView(generics.UpdateAPIView):
         except CustomUser.DoesNotExist as e:
             return Response({'msg': str(e)}, status=status.HTTP_404_NOT_FOUND)
 
+        if request.user not in chat_channel.admins.all():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         chat_channel.members.add(*users)
         chat_channel.save()
 
-        serializer = ChatChannelMembersModifyResponseSerializer(chat_channel)
+        serializer = ChatChannelSerializer(chat_channel)
         return Response(serializer.data)
 
 
-class ChatChannelDeleteMembersView(generics.UpdateAPIView):
+class ChatChannelDeleteMembersView(generics.DestroyAPIView):
     queryset = ChatChannel.objects.all()
     http_method_names = ['delete']
     permission_classes = [permissions.IsAuthenticated]
@@ -158,12 +176,16 @@ class ChatChannelDeleteMembersView(generics.UpdateAPIView):
         """
         특정 유저를 삭제 할 때 쓰는 엔드포인트 입니다.
         한번에 한명만 없앨 수 있습니다.
+        `admins`에 없는 유저는 401을 반환 합니다. (특정 유저를 없앨 수 없습니다.)
         """
         username = kwargs.get('username', None)
-        print(f'{username=}')
         user = get_object_or_404(CustomUser, username__exact=username)
 
         chat_channel: ChatChannel = self.get_queryset()
+
+        if request.user not in chat_channel.admins.all():
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+
         chat_channel.members.remove(user)
         chat_channel.save()
